requirements.txt
# Core ML Framework
torch>=2.0.0
torchvision>=0.15.0

# Data Processing
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0

# Network Analysis
networkx>=3.0

# Visualization
matplotlib>=3.7.0

# Configuration
pyyaml>=6.0

# Utilities
python-dateutil>=2.8.0

# Optional: For production deployments
# gunicorn>=21.0.0
# flask>=2.3.0
# prometheus-client>=0.17.0

